requirements.txt
Spire.Doc==11.10.0
fastapi
PyJWT
passlib
deep_translator
torch
pymongo
uvicorn
fastapi_mail
pdfplumber
transformers
python-docx 
qdrant_client
numpy
sentence_transformers
pdf2image
pytesseract
motor
openai==0.28 
pydantic
python-multipart
llama-index
